+++ src/mongocxx/test/spec/crud.cpp
@@ -30,8 +30,6 @@ void _run_crud_tests_in_file(std::string test_path) {
 }
 
 TEST_CASE("CRUD legacy spec automated tests", "[crud_spec]") {
-    instance::current();
-
     run_tests_in_suite("CRUD_LEGACY_TESTS_PATH", _run_crud_tests_in_file);
 }
 } // namespace