+++ src/mongocxx/test/spec/read_write_concern.cpp
@@ -24,8 +24,6 @@ using namespace mongocxx;
 using namespace spec;
 
 TEST_CASE("Read / Write concern spec tests", "[read_write_concern_spec]") {
-    instance::current();
-
     // Reuse the transactions test runner.
     run_tests_in_suite("READ_WRITE_CONCERN_OPERATION_TESTS_PATH", &run_transactions_tests_in_file);
 }